Did a ton of work on the vines mechanic in a somewhat futile attempt to make it more efficient.
In my mind I thought "Tons of blocks all placed at the same time causing
momentary lag and freezing? Spread it out over as many ticks as possible
so it isn't hitting all at once!" Well, experimentally it seems that
spreading the growth over a longer period of time is just making the lag
last a longer period of time. The placement of the blocks seems much
more responsive when they are all placed as quickly as possible, so the
client environment only has to render a few massive changes to the
environment ranther than continual substantial changes over a longer
period of time.

I think I'm going to move this vines code into it's own class to clean
some of this up, and experiment with a simpler, qucker approach.